+++ kraftwerk-functional-tests/src/test/java/cucumber/functional_tests/StandardVtlDefinitions.java
@@ -3,17 +3,19 @@ package cucumber.functional_tests;
 import com.opencsv.CSVReader;
 import com.opencsv.exceptions.CsvException;
 import cucumber.TestConstants;
-import fr.insee.kraftwerk.core.Constants;;
+import fr.insee.kraftwerk.core.Constants;
 import fr.insee.kraftwerk.core.utils.CsvUtils;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -25,24 +27,23 @@ public class StandardVtlDefinitions {
     @Given("We have test standard vtl scripts")
     public void generate_test_vtl_scripts() throws IOException {
         // Unimode
-        String vtlScript = "TEL := TEL[calc REPORTINGDATA.OUTCOME_SPOTTING := \"TEST\"];\n" +
-                "FAF := FAF[calc TESTVTLFAF := \"TESTFAF\"];";
+        // TEL
+        String vtlScript = "TEL := TEL[calc REPORTINGDATA.OUTCOME_SPOTTING := \"TEST\"];\n";
         Path path = Files.createDirectories(Path.of(Constants.VTL_FOLDER_PATH).resolve("unimode")).resolve("TEL.vtl");
 
         if(!Files.exists(path)) Files.createFile(path);
         Files.write(path,vtlScript.getBytes());
 
-
-        // Reconciliation
-        vtlScript = "MULTIMODE : = MULTIMODE[calc TESTVTLRECONCILIATION := \"TESTRECONCILIATION\"];";
-        path = Files.createDirectories(Path.of(Constants.VTL_FOLDER_PATH).resolve("reconciliation")).resolve("reconciliation.vtl");
+        // FAF
+        vtlScript = "FAF := FAF[calc TESTVTLFAF := \"TESTFAF\"];";
+        path = Files.createDirectories(Path.of(Constants.VTL_FOLDER_PATH).resolve("unimode")).resolve("FAF.vtl");
 
         if(!Files.exists(path)) Files.createFile(path);
         Files.write(path,vtlScript.getBytes());
 
 
         // Multimode
-        vtlScript = "MULTIMODE : = MULTIMODE[calc TESTVTLMULTIMODE := \"TESTMULTIMODE\"];";
+        vtlScript = "MULTIMODE := MULTIMODE[calc TESTVTLMULTIMODE := \"TESTMULTIMODE\"];";
         path = Files.createDirectories(Path.of(Constants.VTL_FOLDER_PATH).resolve("multimode")).resolve("multimode.vtl");
 
         if(!Files.exists(path)) Files.createFile(path);
@@ -50,7 +51,7 @@ public class StandardVtlDefinitions {
 
 
         // Information levels
-        vtlScript = "RACINE : = RACINE[calc TESTVTLINFOLEVEL := \"TESTINFOLEVEL\"];";
+        vtlScript = "MULTIMODE := MULTIMODE[calc TESTVTLINFOLEVEL := \"TESTINFOLEVEL\"];";
         path = Files.createDirectories(Path.of(Constants.VTL_FOLDER_PATH).resolve("information_levels")).resolve("information_levels.vtl");
 
         if(!Files.exists(path)) Files.createFile(path);
@@ -80,11 +81,29 @@ public class StandardVtlDefinitions {
             fieldIndex++;
         }
 
+        //Exclude header
+        content.remove(0);
+
         for (String[] row : content){
             //Row syntax assert
-            assertThat(row).hasSize(fieldIndex + 1);
+            assertThat(row).hasSize(fieldIndex);
             //Row content assert
             assertThat(row[expectedFieldIndex]).isEqualTo(expectedContent);
+        }
+    }
+
+    @Then("we shouldn't have any _keep file in {string} output directory")
+    public void check_lack_keep_file(String surveyName) throws IOException {
+        try(Stream<Path> stream = Files.list(outDirectory.resolve(surveyName))){
+            Set<String> fileNames = stream
+                    .filter(file -> !Files.isDirectory(file))
+                    .map(Path::getFileName)
+                    .map(Path::toString)
+                    .collect(Collectors.toSet());
+
+            for(String fileName : fileNames){
+                assertThat(fileName).doesNotEndWith("_keep.csv");
+            }
         }
     }
 }